requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
wordcloud
joblib
torch
torchvision
Pillow
